+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/RuleSetLoader.java
@@ -21,7 +21,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import net.sourceforge.pmd.PMDConfiguration;
-import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.Language;
 import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.lang.rule.internal.RuleSetReferenceId;
@@ -195,11 +194,10 @@ public final class RuleSetLoader {
      * Loads a list of rulesets, if any has an error, report it on the contextual
      * error reporter instead of aborting, and continue loading the rest.
      *
-     * <p>Internal API: might be published later, or maybe in PMD 7 this
+     * @apiNote Internal API: might be published later, or maybe this
      * will be the default behaviour of every method of this class.
      */
-    @InternalApi
-    public List<RuleSet> loadRuleSetsWithoutException(List<String> rulesetPaths) {
+    List<RuleSet> loadRuleSetsWithoutException(List<String> rulesetPaths) {
         List<RuleSet> ruleSets = new ArrayList<>(rulesetPaths.size());
         boolean anyRules = false;
         boolean error = false;