+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/internal/RuleApplicator.java
@@ -4,6 +4,8 @@
 
 package net.sourceforge.pmd.lang.rule.internal;
 
+import static net.sourceforge.pmd.lang.rule.InternalApiBridge.ruleSetApplies;
+
 import java.util.Collection;
 import java.util.Iterator;
 
@@ -19,7 +21,6 @@ import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.ast.RootNode;
 import net.sourceforge.pmd.lang.rule.Rule;
-import net.sourceforge.pmd.lang.rule.RuleSet;
 import net.sourceforge.pmd.reporting.FileAnalysisListener;
 import net.sourceforge.pmd.reporting.InternalApiBridge;
 import net.sourceforge.pmd.reporting.Report.ProcessingError;
@@ -58,7 +59,7 @@ public class RuleApplicator {
 
     private void applyOnIndex(TreeIndex idx, Collection<? extends Rule> rules, FileAnalysisListener listener) {
         for (Rule rule : rules) {
-            if (!RuleSet.applies(rule, currentLangVer)) {
+            if (!ruleSetApplies(rule, currentLangVer)) {
                 continue; // No point in even trying to apply the rule
             }